requirements.txt
streamlit
plotly
pandas
openpyxl
streamlit_option_menu
python-dotenv



